% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_dbgap_interface.R
\name{manifest_from_gdc}
\alias{manifest_from_gdc}
\title{Query GDC/dbGaP to form an HTML summary}
\usage{
manifest_from_gdc(project.id, output.file = NULL, format.file = NULL)
}
\arguments{
\item{project.id}{A single string value describing a dataset in GDC.}

\item{output.file}{Name of the resulting HTML file.  By default it is based off of the project ID.}

\item{format.file}{A CSV file that overrides the use of the formatting table 
bundled with the package.}
}
\value{
Returns the path to the generated HTML file.
}
\description{
Using the supplied Genomic Data Commons (GDC) project ID, retrieves information about
the dataset including general information, access, authors, composition and basic
information on ethical, legal and social aspects to form an 'Aggregator' Persona BioMedical
Data Manifest HTML file.  NOTE: This functionality is very fragile since the information
it attempts to gather is not consistently well structured.
}
\examples{
\dontrun{manifest_from_gdc("TARGET-AML")}
}
