% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_from_excel.R
\name{manifest_from_excel}
\alias{manifest_from_excel}
\title{Generate an HTML summary from a survey}
\usage{
manifest_from_excel(
  excel.file,
  persona = c("DM/Comp", "Bench/Clinical"),
  output.file = NULL,
  format.file = NULL
)
}
\arguments{
\item{excel.file}{An Excel file containing survey results}

\item{persona}{The type of Persona to use, one of 'DM/Comp' or 'Bench/Clinical'}

\item{output.file}{Name of the HTML file to output.  Defaults to 
'biomedical_data_manifest.html'}

\item{format.file}{A CSV file that overrides the use of the formatting table 
bundled with the package.}
}
\value{
"The path to the generated HTML file"
}
\description{
Given an Excel output from a survey tool such as Microsoft Forms populate
a BioMedical Data Manifest HTML template for a given Persona.  Changing
Personas removes questions/answers according to Bottomly et al In Prep.
}
\examples{
\dontrun{
example.xl <- system.file("extdata", "Biomedical_Data_Manifest_Template_Example.xlsx", package = "BioDataManifest")
manifest_from_excel(example.xl)
}
}
