% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avail_projects}
\alias{avail_projects}
\alias{unnest.expands}
\alias{plot.fun}
\title{Retrieve available project names}
\usage{
avail_projects()

unnest.expands(tbl, expand_col)

\method{plot}{fun}(tbl, plot.col, out.dir = NULL)
}
\arguments{
\item{tbl}{A tibble}

\item{expand_col}{Columns of the tibble to expand}

\item{plot.col}{The tibble column name used to form the plot}

\item{out.dir}{A directory to save the plots to}
}
\value{
character vector of project ids

New tibble from the expansion of the list-column

A character vector indicating the plot path.
}
\description{
Retrieve a character vector of all available project names from the GDC.  These
can be use with the `gdc_datashard` function

Especially when using `GenomicDataCommons::expands` tibbles are returned that
have columns with lists containing nested tibbles.  This function expands those
to standard tibbles for easy merging with the original.

Given a tibble, and the column to plot, create a basic summary figure, either
a bar chart or histogram depending on whether the data is character/factor or
numeric.
}
